make window error remove
